+++ packages/components/src/atoms/divider/styled.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { StyledBoxProps } from '../../primatives';
+import { StyledBoxProps } from '../../primal';
 
 export const StyledHr = styled.hr`
   ${StyledBoxProps}